Migrate to native TypeScript 7 and oxlint
- Bump typescript to native 7, Next to 16.3 canary
- Enable experimental.useTypeScriptCli for native type-checking
- Set tsconfig target es2020 (es5 removed in TS7)
- Replace ESLint with oxlint (no TS JS-API dependency)
- Convert jest.config.ts to .mjs, drop ts-node
- Point editor config at oxc + TS native preview
