--- eslint.config.mjs
@@ -1,25 +1,0 @@
-import { defineConfig } from "eslint/config";
-import nextCoreWebVitals from "eslint-config-next/core-web-vitals";
-import path from "node:path";
-import { fileURLToPath } from "node:url";
-import js from "@eslint/js";
-import { FlatCompat } from "@eslint/eslintrc";
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-const compat = new FlatCompat({
-  baseDirectory: __dirname,
-  recommendedConfig: js.configs.recommended,
-  allConfig: js.configs.all,
-});
-
-export default defineConfig([
-  {
-    extends: [
-      ...nextCoreWebVitals,
-      // TODO: Include Storybook linting after upgrading to latest version
-      // ...compat.extends("plugin:storybook/recommended"),
-      ...compat.extends("prettier"),
-    ],
-  },
-]);
