+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    "target": "es5",
+    "target": "es2020",
     "lib": [
       "dom",
       "dom.iterable",
@@ -17,11 +17,6 @@
     "isolatedModules": true,
     "jsx": "react-jsx",
     "incremental": true,
-    "plugins": [
-      {
-        "name": "next"
-      }
-    ],
     "paths": {
       "~/*": [
         "./app/*"
@@ -35,7 +30,7 @@
       "~logo/*": [
         "./app/components/Logo/*"
       ],
-       "~utils/*": [
+      "~utils/*": [
         "./app/utils/*"
       ],
       "~assets/*": [
@@ -44,6 +39,11 @@
     },
     "types": [
       "@testing-library/jest-dom/jest-globals"
+    ],
+    "plugins": [
+      {
+        "name": "next"
+      }
     ]
   },
   "include": [
